Add hex version badge to the Readme
Hi!

This is a cool add - link to your Hex package, in a form of SVG badge aka shield. As Hex package gives a link to here (that's how I found this project), nice to give a link back. Also it visually indicates that your project is packaged, which is a Good Thing (tm). :)

More shields related to Hex package - see http://shields.io/

hope this helps,

W.